# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "FastAPI",
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/buckets": {
			"get": {
				"summary": "List Buckets",
				"operationId": "list_buckets_api_v1_buckets_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Create Bucket",
				"operationId": "create_bucket_api_v1_buckets_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBucket"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/buckets/{bucket_name}": {
			"delete": {
				"summary": "Delete Bucket",
				"operationId": "delete_bucket_api_v1_buckets__bucket_name__delete",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/buckets/{bucket_name}/objects": {
			"post": {
				"summary": "Upload Object",
				"operationId": "upload_object_api_v1_buckets__bucket_name__objects_post",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UploadObject"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"get": {
				"summary": "List Objects",
				"operationId": "list_objects_api_v1_buckets__bucket_name__objects_get",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					},
					{
						"name": "prefix",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"default": "",
							"title": "Prefix"
						}
					},
					{
						"name": "max_keys",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"default": 1000,
							"title": "Max Keys"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/buckets/{bucket_name}/objects/{key}": {
			"delete": {
				"summary": "Delete Object",
				"operationId": "delete_object_api_v1_buckets__bucket_name__objects__key__delete",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					},
					{
						"name": "key",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Key"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/buckets/{bucket_name}/website": {
			"post": {
				"summary": "Configure Website Hosting",
				"operationId": "configure_website_hosting_api_v1_buckets__bucket_name__website_post",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBucketWebsiteHosting"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/buckets/{bucket_name}/policy": {
			"post": {
				"summary": "Set Bucket Policy",
				"operationId": "set_bucket_policy_api_v1_buckets__bucket_name__policy_post",
				"parameters": [
					{
						"name": "bucket_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Bucket Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBucketPolicy"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateBucket": {
				"properties": {
					"bucket": {
						"type": "string",
						"title": "Bucket"
					},
					"region": {
						"type": "string",
						"enum": [
							"EU",
							"af-south-1",
							"ap-east-1",
							"ap-northeast-1",
							"ap-northeast-2",
							"ap-northeast-3",
							"ap-south-1",
							"ap-south-2",
							"ap-southeast-1",
							"ap-southeast-2",
							"ap-southeast-3",
							"ap-southeast-4",
							"ap-southeast-5",
							"ca-central-1",
							"cn-north-1",
							"cn-northwest-1",
							"eu-central-1",
							"eu-north-1",
							"eu-south-1",
							"eu-west-1",
							"eu-west-2",
							"eu-west-3",
							"me-central-1",
							"me-south-1",
							"sa-east-1",
							"us-east-1",
							"us-east-2",
							"us-west-1",
							"us-west-2"
						],
						"title": "Region"
					},
					"acl": {
						"type": "string",
						"enum": [
							"authenticated-read",
							"private",
							"public-read",
							"public-read-write"
						],
						"title": "Acl"
					},
					"versioning": {
						"type": "boolean",
						"title": "Versioning",
						"default": false
					},
					"encryption": {
						"type": "boolean",
						"title": "Encryption",
						"default": false
					},
					"logging": {
						"type": "boolean",
						"title": "Logging",
						"default": false
					}
				},
				"type": "object",
				"required": [
					"bucket",
					"region",
					"acl"
				],
				"title": "CreateBucket"
			},
...
			"CreateBucketPolicy": {
				"properties": {
					"bucket": {
						"type": "string",
						"title": "Bucket"
					},
					"policy": {
						"type": "string",
						"title": "Policy"
					}
				},
				"type": "object",
				"required": [
					"bucket",
					"policy"
				],
				"title": "CreateBucketPolicy"
			},
			"CreateBucketWebsiteHosting": {
				"properties": {
					"bucket": {
						"type": "string",
						"title": "Bucket"
					},
					"index_document": {
						"type": "string",
						"title": "Index Document",
						"default": "index.html"
					},
					"error_document": {
						"type": "string",
						"title": "Error Document",
						"default": "error.html"
					}
				},
				"type": "object",
				"required": [
					"bucket"
				],
				"title": "CreateBucketWebsiteHosting"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"UploadObject": {
				"properties": {
					"bucket": {
						"type": "string",
						"title": "Bucket"
					},
					"key": {
						"type": "string",
						"title": "Key"
					},
					"file_path": {
						"type": "string",
						"title": "File Path"
					},
					"content_type": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Content Type"
					},
					"metadata": {
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"title": "Metadata"
					}
				},
				"type": "object",
				"required": [
					"bucket",
					"key",
					"file_path"
				],
				"title": "UploadObject"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError"
			},
...
		}
	}
}